Refactor WebLab selenium tests
